fix: remove unneccesary code and refactor set value condition
