+++ src/app/app.ts
@@ -1,6 +1,4 @@
-// import { ProgressBlock } from './components/progressBlock/progressBlock';
-// import type BaseComponent from './components/baseComponent/baseComponent';
-import progressBlock from '@components/progressBlock/createProgressBlock';
+import progressBlock from '@components/progressBlock/progressBlock';
 
 class App {
   constructor(private progressBlock: HTMLElement) {}
@@ -16,6 +14,5 @@ if (!root) {
   throw new Error('Root element not found');
 }
 
-// const app = new App(ProgressBlock());
 const app = new App(progressBlock.element);
 app.createApp(root);
