--- src/app/components/progressBlock/_progressBlock.module.scss
@@ -1,8 +1,0 @@
-.progress-block {
-  display: flex;
-  flex-direction: column;
-}
-
-.hide {
-  display: none;
-}
