--- src/app/components/progressBlock/controlComponent/_controlComponent.module.scss
@@ -1,5 +1,0 @@
-.input-group {
-  display: flex;
-  flex-direction: row-reverse;
-  gap: 22px;
-}
