--- src/app/components/progressBlock/controlPanel/_controlPanel.module.scss
@@ -1,1 +1,0 @@
-// styles
