--- src/app/components/progressBlock/progressBlockBody/_progressBlockBody.module.scss
@@ -1,4 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-}
