--- src/app/components/progressBlock/progressBlockBody/progressBlockBody.ts
@@ -1,16 +1,0 @@
-import BaseComponent from '@components/baseComponent/baseComponent';
-import styles from './_progressBlockBody.module.scss';
-import ControlPanelComponent, { ControlPanel } from '../controlPanel/controlPanel';
-
-export default class ProgressBlockBodyComponent extends BaseComponent<'main'> {
-  constructor(private controlPanel: ControlPanelComponent) {
-    super({ tag: 'main', classes: [styles.main] });
-    this.render();
-  }
-
-  private render(): void {
-    this.append([this.controlPanel.getElement()]);
-  }
-}
-
-export const ProgressBlockBody = (): ProgressBlockBodyComponent => new ProgressBlockBodyComponent(ControlPanel());
