--- src/app/components/progressBlock/progressBlockHeader/_header.module.scss
@@ -1,7 +1,0 @@
-.header {
-  background-color: white;
-}
-
-.title {
-  font-size: 1rem;
-}
