+++ src/app/components/progressBlock/progressBlockView/canvasComponent/_canvas.module.scss
