+++ src/app/components/progressBlock/progressBlockView/canvasComponent/canvasComponent.ts
