+++ src/app/components/progressBlock/progressBlockView/progressCircle/progressCircle.ts
